# api/generated/openapi/com/mts/online_shop/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mts.online_shop.api;

import com.mts.online_shop.model.LoginRequest;
import com.mts.online_shop.model.LoginResponse;
import com.mts.online_shop.model.MessageResponse;
import com.mts.online_shop.model.RegisterRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-10T20:48:52.004416800+03:00[Europe/Moscow]")
@Validated
@Tag(name = "Auth", description = "Аутентификация и регистрация")
public interface AuthApi {

    /**
     * POST /auth/login : Вход (логин и пароль)
     *
     * @param loginRequest  (required)
     * @return Успешный вход, возвращается JWT (status code 200)
     */
    @Operation(
        operationId = "login",
        summary = "Вход (логин и пароль)",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Успешный вход, возвращается JWT", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LoginResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<LoginResponse> login(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    );


    /**
     * POST /auth/register : Регистрация нового пользователя
     *
     * @param registerRequest  (required)
     * @return Пользователь создан (status code 201)
     */
    @Operation(
        operationId = "register",
        summary = "Регистрация нового пользователя",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Пользователь создан", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/register",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<MessageResponse> register(
        @Parameter(name = "RegisterRequest", description = "", required = true) @Valid @RequestBody RegisterRequest registerRequest
    );

}