# api/generated/openapi/com/mts/online_shop/api/CartApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mts.online_shop.api;

import com.mts.online_shop.model.AddCartItemRequest;
import com.mts.online_shop.model.CartResponse;
import com.mts.online_shop.model.MessageResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-10T20:48:52.004416800+03:00[Europe/Moscow]")
@Validated
@Tag(name = "Cart", description = "Корзина текущего пользователя")
public interface CartApi {

    /**
     * POST /cart/items : Добавить товар в корзину
     *
     * @param addCartItemRequest  (required)
     * @return Товар добавлен в корзину (status code 201)
     */
    @Operation(
        operationId = "addCartItem",
        summary = "Добавить товар в корзину",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Товар добавлен в корзину", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/cart/items",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<MessageResponse> addCartItem(
        @Parameter(name = "AddCartItemRequest", description = "", required = true) @Valid @RequestBody AddCartItemRequest addCartItemRequest
    );


    /**
     * DELETE /cart/clear : Очистить корзину
     *
     * @return Корзина очищена (status code 200)
     */
    @Operation(
        operationId = "clearCart",
        summary = "Очистить корзину",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Корзина очищена", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/cart/clear",
        produces = { "application/json" }
    )
    ResponseEntity<MessageResponse> clearCart(
        
    );


    /**
     * GET /cart : Получить содержимое корзины
     *
     * @return Корзина пользователя (status code 200)
     */
    @Operation(
        operationId = "getCart",
        summary = "Получить содержимое корзины",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Корзина пользователя", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CartResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/cart",
        produces = { "application/json" }
    )
    ResponseEntity<CartResponse> getCart(
        
    );


    /**
     * DELETE /cart/items/{itemId} : Удалить товар из корзины
     *
     * @param itemId  (required)
     * @return Товар удален из корзины (status code 200)
     */
    @Operation(
        operationId = "removeCartItem",
        summary = "Удалить товар из корзины",
        tags = { "Cart" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Товар удален из корзины", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/cart/items/{itemId}",
        produces = { "application/json" }
    )
    ResponseEntity<MessageResponse> removeCartItem(
        @Parameter(name = "itemId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("itemId") Long itemId
    );

}