# api/generated/openapi/com/mts/online_shop/api/OrdersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mts.online_shop.api;

import com.mts.online_shop.model.MessageResponse;
import com.mts.online_shop.model.OrderListResponse;
import com.mts.online_shop.model.OrderResponse;
import com.mts.online_shop.model.PaymentRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-10T20:48:52.004416800+03:00[Europe/Moscow]")
@Validated
@Tag(name = "Orders", description = "Заказы и оплата")
public interface OrdersApi {

    /**
     * POST /orders/{orderId}/cancel : Отменить заказ с возвратом денег
     *
     * @param orderId  (required)
     * @return Заказ отменен, деньги возвращены (status code 200)
     */
    @Operation(
        operationId = "cancelOrder",
        summary = "Отменить заказ с возвратом денег",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Заказ отменен, деньги возвращены", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MessageResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/orders/{orderId}/cancel",
        produces = { "application/json" }
    )
    ResponseEntity<MessageResponse> cancelOrder(
        @Parameter(name = "orderId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("orderId") Long orderId
    );


    /**
     * POST /orders/create : Создать заказ с оплатой (транзакция)
     *
     * @param paymentRequest  (required)
     * @return Заказ создан и оплачен (status code 201)
     *         or Ошибка валидации данных карты или пустая корзина (status code 400)
     *         or Внутренняя ошибка сервера (status code 500)
     */
    @Operation(
        operationId = "createOrder",
        summary = "Создать заказ с оплатой (транзакция)",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Заказ создан и оплачен", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Ошибка валидации данных карты или пустая корзина"),
            @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/orders/create",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<OrderResponse> createOrder(
        @Parameter(name = "PaymentRequest", description = "", required = true) @Valid @RequestBody PaymentRequest paymentRequest
    );


    /**
     * GET /orders : Получить заказы текущего пользователя
     *
     * @return Список заказов (status code 200)
     */
    @Operation(
        operationId = "getOrders",
        summary = "Получить заказы текущего пользователя",
        tags = { "Orders" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Список заказов", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderListResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/orders",
        produces = { "application/json" }
    )
    ResponseEntity<OrderListResponse> getOrders(
        
    );

}