# api/generated/openapi/com/mts/online_shop/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mts.online_shop.api;

import com.mts.online_shop.model.Product;
import com.mts.online_shop.model.ProductListResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-10T20:48:52.004416800+03:00[Europe/Moscow]")
@Validated
@Tag(name = "Products", description = "Каталог товаров")
public interface ProductsApi {

    /**
     * GET /products/{id} : Получить информацию о товаре
     *
     * @param id  (required)
     * @return Информация о товаре (status code 200)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Получить информацию о товаре",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Информация о товаре", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    ResponseEntity<Product> getProductById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /products : Список всех доступных товаров
     *
     * @return Список товаров (status code 200)
     */
    @Operation(
        operationId = "getProducts",
        summary = "Список всех доступных товаров",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Список товаров", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductListResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products",
        produces = { "application/json" }
    )
    ResponseEntity<ProductListResponse> getProducts(
        
    );

}