# api/openapi.yml
openapi: 3.0.3
info:
  title: MTS Online Shop
  version: 0.1.0
  description: REST API интернет-магазина
servers:
  - url: http://localhost:8080/api
    description: Dev Server
...
paths:
  /auth/login:
    post:
      summary: Вход (логин и пароль)
      tags: [Auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход, возвращается JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /auth/register:
    post:
      summary: Регистрация нового пользователя
      tags: [Auth]
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /products/{id}:
    get:
      summary: Получить информацию о товаре
      tags: [Products]
      operationId: getProductById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Информация о товаре
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /cart:
    get:
      summary: Получить содержимое корзины
      tags: [Cart]
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
  /cart/items:
    post:
      summary: Добавить товар в корзину
      tags: [Cart]
      operationId: addCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "201":
          description: Товар добавлен в корзину
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/items/{itemId}:
    delete:
      summary: Удалить товар из корзины
      tags: [Cart]
      operationId: removeCartItem
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Товар удален из корзины
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/clear:
    delete:
      summary: Очистить корзину
      tags: [Cart]
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Корзина очищена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /orders:
    get:
      summary: Получить заказы текущего пользователя
      tags: [Orders]
      operationId: getOrders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /orders/create:
    post:
      summary: Создать заказ с оплатой (транзакция)
      tags: [Orders]
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
            examples:
              payment_example:
                summary: "Example payment data"
                value:
                  cardNumber: "1234123412341234"
                  cvv: "444"
                  expiresAt: "09/30"
      responses:
        "201":
          description: Заказ создан и оплачен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Ошибка валидации данных карты или пустая корзина
        "500":
          description: Внутренняя ошибка сервера
  /orders/{orderId}/cancel:
    post:
      summary: Отменить заказ с возвратом денег
      tags: [Orders]
      operationId: cancelOrder
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Заказ отменен, деньги возвращены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
                
  # Admin APIs
...
components:
  schemas:
    LoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          example: "user1"
        password:
          type: string
          example: "password123"
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    RegisterRequest:
      type: object
      required:
        - login
        - email
        - password
        - name
      properties:
        login:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Операция выполнена успешно"
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Пакет интернета 5Gb"
        price:
          type: number
          format: double
          example: 199.99
...
    AddCartItemRequest:
      type: object
      required:
        - productId
      properties:
        productId:
          type: integer
          format: int64
          example: 1
    CartItem:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
          example: 123
        product:
          $ref: "#/components/schemas/Product"
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 2
        totalPrice:
          type: number
          format: double
          example: 599.97
        status:
          type: string
          enum: [CREATED, PAID, DELIVERED, CANCELLED]
          example: "PAID"
        createdAt:
          type: string
          format: date-time
          example: "2024-03-15T10:30:00Z"
    OrderResponse:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CartItem"
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    PaymentRequest:
      type: object
      required:
        - cardNumber
        - expiresAt
        - cvv
      properties:
        cardNumber:
          type: string
          example: "1234123412341234"
        expiresAt:
          type: string
          example: "12/30"
        cvv:
          type: string
          example: "111"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT