# backend/src/main/java/com/mts/online_shop/model/CartItem.java
package com.mts.online_shop.model;

public class CartItem {
    private Long id;
    private Integer quantity;
    private Product product;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Integer getQuantity() {
        return quantity;
    }

    public void setQuantity(Integer quantity) {
        this.quantity = quantity;
    }

    public Product getProduct() {
        return product;
    }

    public void setProduct(Product product) {
        this.product = product;
    }
}